+++ packages/dnr-rulesets/test/tasks/local-scripts/local-script-invalid-expected.js
@@ -32,5 +32,6 @@ export const localScriptRules = {
         } catch (error) {
             console.error('Error executing AG js rule with uniqueId "test-id" due to: ' + error);
         }
-    }
+    },
+    __PLACEHOLDER__: () => {}
 };