+++ packages/examples/adguard-api/extension/src/index.ts
@@ -18,7 +18,15 @@ import {
         ],
         filteringEnabled: true,
         allowlist: ['www.example.com'],
-        rules: ['example.org##h1'],
+        /* eslint-disable @typescript-eslint/quotes */
+        rules: [
+            'example.org##h1',
+            `#%#console.log('generic script injected at: ', Date.now());`,
+            `#%#//scriptlet('log', 'generic scriptlet injected')`,
+            `example.net#%#console.log('specific script injected at: ', Date.now());`,
+            `example.net#%#//scriptlet('log', 'specific scriptlet injected')`,
+        ],
+        /* eslint-enable @typescript-eslint/quotes */
         filterRulesUrl: 'https://filters.adtidy.org/extension/chromium/filters/{filter_id}.txt',
         filtersMetadataUrl: 'https://filters.adtidy.org/extension/chromium/filters.json',
     };