+++ packages/tswebextension/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.17] - 2025-12-01
+
+### Changed
+
+- Export `LocalScriptRules` type from MV2.
+
+[3.2.17]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v3.2.17
+
 ## [3.2.16] - 2025-11-26
 
 ### Changed
@@ -13,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated [@adguard/agtree] to `v3.3.1`.
 - Updated [@adguard/scriptlets] to `v2.2.13`.
 - Updated [@adguard/tsurlfilter] to `v3.5.1`.
+
+[3.2.16]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v3.2.16
 
 ## [3.2.15] - 2025-11-12